modify code

+++ back-end/internal/repository/server.go
@@ -6,7 +6,7 @@ import (
 )
 
 type ServerRepository interface {
-	GetServersFiltered(perPage, offset int, status, field, order string) ([]model.Server, int64, error)
+	GetServersFiltered(perPage, offset int, status, field, order string) ([]model.Server, int, error)
 	GetServersByOptionalDateRange(startCreated, endCreated, startUpdated, endUpdated time.Time, field, order string) ([]model.Server, error)
 	Create(s *model.Server) error
 	CreateMany(servers []model.Server) error
